Get the value of DisableNetork via proper torrc parsing
@adrelanos suggested in the [pull request](https://github.com/Whonix/anon-shared-helper-scripts/pull/3#issuecomment-377913469):

> If you would like to improve this implementation, could you please
  implement "the real thing" (tm)? I.e. "parsing as Tor does",
  i.e. starting from tor-service-defaults, /etc/tor/torrc and then
  parsing and following %include as written in Tor configs?

Since "the real thing" has been implemented as part of the
anon-verify, we simple take the advantege of it. Note that a PR to
anon-verify need to be approved before the code here can work as
expected.

Also note that this solution should be once-for-all. If it really
requires any maintainnace in the future, the maintainnace should be
done in torrc-parser as part of the anon-gw-anonymizer-config package.