Remove extra_colors option
Exclude SGR is much more reliable now. SGR bit modes 16 and 24 bit were
difficult to generate the regex, but as it became simpler, separating
them from 4-bit SGR does not make sense. Although it was an optional
parameter, it now makes the code easier to read with less options per
function.

Also change the word color to SGR when appropriate, as it can show
graphics that are not colors such as italic, underline, blink, bold etc.